packages/backend-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-browser-backend-python"
version = "1.0.0"
description = "Python/FastAPI backend for Agentic Browser Plugin"
authors = [{name = "Agentic Browser Team"}]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "playwright>=1.40.0",
    "langchain>=0.0.350",
    "langchain-community>=0.0.1",
    "huggingface-hub>=0.19.4",
    "transformers>=4.36.2",
    "torch>=2.1.2",
    "sqlalchemy>=2.0.23",
    "sqlmodel>=0.0.14",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.2",
    "httpx>=0.25.2",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]
requires-python = ">=3.9"

[project.scripts]
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"